Refactor invitation routes into dedicated router
